[REFEAT] Altera metodo delete do controller Cliente

+++ src/main/java/com/example/barguel/controllers/ClienteController.java
@@ -1,7 +1,9 @@
 package com.example.barguel.controllers;
 
 import com.example.barguel.dtos.ClienteDto;
+import com.example.barguel.models.aluguel.AluguelModel;
 import com.example.barguel.models.cliente.ClienteModel;
+import com.example.barguel.services.AluguelService;
 import com.example.barguel.services.ClienteService;
 import org.springframework.beans.BeanUtils;
 import org.springframework.http.HttpStatus;
@@ -9,9 +11,7 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpSession;
 import javax.validation.Valid;
-import java.net.URI;
 import java.util.List;
 import java.util.Optional;
 import java.util.UUID;
@@ -22,9 +22,11 @@ import java.util.UUID;
 public class ClienteController {
 
     final ClienteService clienteService;
+    final AluguelService aluguelService;
 
-    public ClienteController(ClienteService clienteService) {
+    public ClienteController(ClienteService clienteService, AluguelService aluguelService) {
         this.clienteService = clienteService;
+        this.aluguelService = aluguelService;
     }
     @PostMapping(value = "/save")
     public ResponseEntity<Object> saveCliente(@Valid @RequestBody ClienteDto clienteDto){
@@ -59,9 +61,17 @@ public class ClienteController {
 
             return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Cliente não encontrado!");
         }
-        //verificar se o cliente possui um aluguel em aberto antes de escluir
-        clienteService.deleteByid(clienteModelOptional.get().getId());
-        return ResponseEntity.status(HttpStatus.OK).body("Cliente deletado com sucesso!");
+        List<AluguelModel> alugueis =aluguelService.findAll()
+                .stream()
+                .filter(aluguel -> aluguel.getCliente().getId().equals(clienteModelOptional.get().getId()))
+                .toList();
+        if(alugueis.isEmpty()){
+            clienteService.deleteByid(clienteModelOptional.get().getId());
+            return ResponseEntity.status(HttpStatus.OK).body("Cliente deletado com sucesso!");
+        }
+        else
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("O cliente não pode ser deletado pois possui um aluguel em aberto!");
+
     }
     @GetMapping(value = "getAll")
     public ResponseEntity<List<ClienteModel>> getAllClientes(HttpServletRequest request){
